[release] refactor: remove code noise
