+++ components/restreamer/src/ffmpeg/mod.rs
@@ -8,7 +8,6 @@ mod restreamer;
 mod restreamer_kind;
 mod restreamers_pool;
 mod transcoding_restreamer;
-mod util;
 
 pub use self::{
     copy_restreamer::CopyRestreamer,
